Fix styles in TabsContainer
