+++ src/pages/Invocation/InvocationPage.tsx
@@ -1,4 +1,4 @@
-import { AlertCircle, Loader } from 'lucide-react';
+import { Loader } from 'lucide-react';
 import React from 'react';
 import { Navigate, useParams } from 'react-router-dom';
 
@@ -7,33 +7,9 @@ import useInvocation from './useInvocation';
 import { useInvocationQuery } from '@/common/api/invocations';
 import Breadcrumb from '@/common/components/Breadcrumb/Breadcrumb';
 import ContractInput from '@/common/components/Input/ContractInput';
-import AuthorizationTab from '@/common/components/Tabs/AuthorizationTab/AuthorizationTab';
-import EventsTab from '@/common/components/Tabs/EventsTab/EventsTab';
-import FunctionsTab from '@/common/components/Tabs/FunctionsTab/FunctionsTab';
-import PreInvocationTab from '@/common/components/Tabs/PreInvocationTab/PreInvocationTab';
-import TestsTab from '@/common/components/Tabs/TestsTab/TestsTab';
+import TabsContainer from '@/common/components/Tabs/TabsContainer';
 import Terminal from '@/common/components/ui/Terminal';
-import {
-	Tabs,
-	TabsContent,
-	TabsList,
-	TabsTrigger,
-} from '@/common/components/ui/tabs';
-import {
-	Tooltip,
-	TooltipContent,
-	TooltipTrigger,
-} from '@/common/components/ui/tooltip';
-import useEditor from '@/common/hooks/useEditor';
 import { Invocation } from '@/common/types/invocation';
-
-const tabs: Record<string, string> = {
-	functions: 'Functions',
-	authorization: 'Authorization',
-	preInvocateScript: 'Pre-invocate script',
-	tests: 'Tests',
-	events: 'Events tracker',
-};
 
 export type InvocationForm = {
 	contractId?: string | null;
@@ -58,11 +34,6 @@ const InvocationPageContent = ({ data }: { data: Invocation }) => {
 		return data.postInvocation ?? '';
 	}, [data]);
 
-	const { setPreInvocationValue, setPostInvocationValue } = useEditor(
-		preInvocationValue,
-		postInvocationValue,
-	);
-
 	const isMissingKeys = React.useMemo(() => {
 		return !data.publicKey || !data.secretKey;
 	}, [data.publicKey, data.secretKey]);
@@ -84,76 +55,12 @@ const InvocationPageContent = ({ data }: { data: Invocation }) => {
 				runInvocation={handleRunInvocation}
 				loading={isLoadingContract || isRunningInvocation}
 			/>
-			<Tabs
-				defaultValue="functions"
-				className="flex-1"
-				data-test="tabs-container"
-			>
-				<TabsList className="" data-test="tabs-list-container">
-					{Object.keys(tabs).map((tab) => {
-						if (tab === 'authorization' && isMissingKeys) {
-							return (
-								<Tooltip delayDuration={0} key={tab}>
-									<TooltipTrigger asChild>
-										<TabsTrigger
-											value={tab}
-											data-test={`functions-tabs-${tab}`}
-										>
-											{tabs[tab]}
-											<AlertCircle className="text-red-500 ml-2" size={16} />
-										</TabsTrigger>
-									</TooltipTrigger>
-									<TooltipContent>
-										<p>Missing keys</p>
-									</TooltipContent>
-								</Tooltip>
-							);
-						}
-
-						return (
-							<TabsTrigger
-								key={tab}
-								value={tab}
-								data-test={`functions-tabs-${tab}`}
-							>
-								{tabs[tab]}
-							</TabsTrigger>
-						);
-					})}
-				</TabsList>
-				<TabsContent value="functions">
-					<FunctionsTab
-						invocationId={data.id}
-						methods={data.methods}
-						selectedMethod={data.selectedMethod}
-					/>
-				</TabsContent>
-				<TabsContent value="authorization">
-					<AuthorizationTab
-						invocationId={data.id}
-						network={data.network}
-						defaultValues={{
-							secretKey: data.secretKey,
-							publicKey: data.publicKey,
-						}}
-					/>
-				</TabsContent>
-				<TabsContent value="preInvocateScript" className="h-[500px]">
-					<PreInvocationTab
-						preInvocationValue={preInvocationValue}
-						setPreInvocationValu={setPreInvocationValue}
-					/>
-				</TabsContent>
-				<TabsContent value="tests" className="h-[500px]">
-					<TestsTab
-						testsInvocationValue={postInvocationValue}
-						setTestsInvocationValue={setPostInvocationValue}
-					/>
-				</TabsContent>
-				<TabsContent value="events">
-					<EventsTab />
-				</TabsContent>
-			</Tabs>
+			<TabsContainer
+				data={data}
+				preInvocationValue={preInvocationValue}
+				postInvocationValue={postInvocationValue}
+				isMissingKeys={isMissingKeys}
+			/>
 			<Terminal entries={contractResponses} />
 		</div>
 	);
